+++ UVtools.Core/Gerber/Apertures/Aperture.cs
@@ -64,18 +64,24 @@ public abstract class Aperture
             case "C":
             {
                 if (match.Groups.Count < 4) return null;
-                if (!double.TryParse(match.Groups[3].Value, NumberStyles.Float, CultureInfo.InvariantCulture, out var diameter)) return null;
-                return new CircleAperture(document, index, diameter);
+                var split = match.Groups[3].Value.Split('X', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
+                if (split.Length == 0) return null;
+                if (!double.TryParse(split[0], NumberStyles.Float, CultureInfo.InvariantCulture, out var diameter)) return null;
+                var holeDiameter = 0.0;
+                if (split.Length > 1 && !double.TryParse(split[1], NumberStyles.Float, CultureInfo.InvariantCulture, out holeDiameter)) return null;
+                return new CircleAperture(document, index, diameter, holeDiameter);
             }
-            case "O":
+            case "O": // OBround
             {
                 if (match.Groups.Count < 4) return null;
                 var split = match.Groups[3].Value.Split('X', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries);
                 if (split.Length < 2) return null;
                 if (!float.TryParse(split[0], NumberStyles.Float, CultureInfo.InvariantCulture, out var width)) return null;
                 if (!float.TryParse(split[1], NumberStyles.Float, CultureInfo.InvariantCulture, out var height)) return null;
+                var holeDiameter = 0.0;
+                if (split.Length > 2 && !double.TryParse(split[2], NumberStyles.Float, CultureInfo.InvariantCulture, out holeDiameter)) return null;
 
-                return new EllipseAperture(document, index, width, height);
+                return new ObroundAperture(document, index, width, height, holeDiameter);
             }
             case "R":
             {
@@ -84,8 +90,10 @@ public abstract class Aperture
                 if (split.Length < 2) return null;
                 if (!float.TryParse(split[0], NumberStyles.Float, CultureInfo.InvariantCulture, out var width)) return null;
                 if (!float.TryParse(split[1], NumberStyles.Float, CultureInfo.InvariantCulture, out var height)) return null;
+                var holeDiameter = 0.0;
+                if (split.Length > 2 && !double.TryParse(split[2], NumberStyles.Float, CultureInfo.InvariantCulture, out holeDiameter)) return null;
 
-                return new RectangleAperture(document, index, width, height);
+                    return new RectangleAperture(document, index, width, height, holeDiameter);
                 }
             case "P":
             {
@@ -94,8 +102,12 @@ public abstract class Aperture
                 if (split.Length < 2) return null;
                 if (!double.TryParse(split[0], NumberStyles.Float, CultureInfo.InvariantCulture, out var diameter)) return null;
                 if (!ushort.TryParse(split[1], NumberStyles.Float, CultureInfo.InvariantCulture, out var vertices)) return null;
+                var rotation = 0.0;
+                if (split.Length > 2 && !double.TryParse(split[2], NumberStyles.Float, CultureInfo.InvariantCulture, out rotation)) return null;
+                var holeDiameter = 0.0;
+                if (split.Length > 3 && !double.TryParse(split[3], NumberStyles.Float, CultureInfo.InvariantCulture, out holeDiameter)) return null;
 
-                return new PolygonAperture(document, index, diameter, vertices);
+                return new PolygonAperture(document, index, diameter, vertices, rotation, holeDiameter);
             }
             default: // macro
             {
@@ -104,7 +116,7 @@ public abstract class Aperture
                 //var parseLine = line.TrimEnd('%', '*');
                 //var commaIndex = parseLine.IndexOf(',')+1;
                 //parseLine = parseLine[commaIndex..];
-                string[] args = { "0" };
+                string[] args = ["0"];
                 if (match.Groups.Count >= 4)
                 {
                     args = args.Concat(match.Groups[3].Value.Split('X', StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)).ToArray();