+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 08/01/2025 - v5.0.4
+
+- PCB Exposure:
+  - (Fix) Polygon primitive vertex count not parsing correctly when having argument (#976)
+  - (Fix) Obround aperture to follow the correct implementation (two semicircles connected by parallel lines tangent to their endpoints) (#976)
+  - (Fix) Implement the "hole diameter" argument in all apertures (#976)
+  - (Fix) Implement the "rotation" argument for the polygon aperture
+
 ## 28/12/2024 - v5.0.3
 
 - Anycubic file format: